Desktop/noor-ana/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client==1.8.0
google-generativeai==0.3.2
redis==5.0.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
better-auth==0.2.0
google-cloud-translate==3.14.0
aiohttp==3.9.0
alembic==1.13.1